Update CSS.md Animation section
Updated the order in which properties appear in the Properties table to match the order in which they should be listed in the Shorthand form. Also added 2 missing properties ( animation-fill-mode & animation-play-state ) and an example for them. See link for more info: https://www.w3schools.com/cssref/css3_pr_animation.asp.