Add cheatsheet on handling generated content with Bulma.
In cases where WYSIWYG editors generate the HTML content of the page,
the content does not have Bulma classes applied to them, which means
they'll be unstyled because that's just how Bulma works. An `<h1>` tag
does not have a style until you add the class `title` to it.

To handle this case, wrap the generated content with `.content` class.