Correct description of ":=" and "=" operators
The description of the ":=" and "=" assignment operators were flipped. As documented in the GNU make manual (https://www.gnu.org/software/make/manual/make.html#Reading-Makefiles), the ":=" assignment operatory is the immediate assignment operation and "=" is the deferred assignment operation. This change switches the descriptions to correspond to the correct operator and augments the description to clarify the behaviour.