gitpod: deploy
Disclaimer: i'm not gitpod staff i just like their product

Proposing to integrate gitpod with this repository which is online IDE based on Theia which allows contributors to open an online environment without the need to fetch deps on their local system.

This service is free for open-source repositories.

Example of the runtime:
https://gitpod.io/#https://github.com/RXT067/cheatsheets/tree/gitpod

Signed-off-by: Jacob Hrbek <kreyren@rixotstudio.cz>