Fix aria-hidden elements
aria-hidden (not area-hidden, as shown in the removed comment) is for accessibility purposes. Sometimes libraries add two associated elements, one for screen readers for disabled people and another one for the display. To avoid automatic detection of the redundant display item in screen readers, one can add aria-hidden to the display element. 

The removed lines of code were messing with many accessible libraries and should be removed.