Fixed bug in WaitGroup example
The waitgroup that is defined in main() with var wg sync.WaitGroup, only has scope inside the main() function. The doOperation() function will not know about it and so the defer wg.Done() line will result in a compile time error. The waitgroup needs to be passed as a parameter to the doOperation() function.