Add missing syntax highlighting to python.md (#2154)
Fixes missing code highlighting in half of the Python code snippets.

The top half of the page had code examples formatted as code blocks with
4 space character indentations and no syntax highlighting as a result.

The bottom half on the other hand, used code blocks surrounded by
"```py" formatting brackets which hints to many markdown renders that
syntax highlighting should be applied.

This unifies the code blocks in the top half of the page to use the same
formatting syntax used in the bottom half with syntax highlighting.

This change was prompted by requests to add the syntax highlighting in
comments on the https://devhints.io/python page.